pyproject.toml
[project]
name = "ai-voice-order-confirmation"
version = "0.1.0"
description = "AI Voice Operating System for Delivery Agent Order Confirmation"
requires-python = ">=3.10"
dependencies = [
    "redis>=5.0.0",
    "psycopg2-binary>=2.9.9",
    "boto3>=1.34.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "websockets>=12.0",
    "hypothesis>=6.98.0",
    "sqlalchemy>=2.0.25",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "celery>=5.3.6",
    "librosa>=0.10.1",
    "webrtcvad>=2.0.10",
    "google-cloud-speech>=2.24.0",
    "google-cloud-texttospeech>=2.16.0",
    "openai>=1.10.0",
    "twilio>=8.11.0",
    "python-statemachine>=2.1.2",
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
